A heap of changes!
I have decided to split the calculation in days and am breaking the data frame into dictionries with the key as the day's date and
the values as the day's profile.

I then cloned this dict and will iterate through the cloned dict performing the optimisation.

I still need to deal with edge cases such as if there is not enough data in a day.

I also think this code needs a heap of optimisation and organisation.